backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
python-dotenv
psycopg2-binary
